fix(Barrage): demo拆解与规范
