+++ src/packages/barrage/doc.zh-TW.md
@@ -16,52 +16,7 @@ import { Barrage } from '@nutui/nutui-react'
 
 :::demo
 
-```tsx
-import React, { useRef } from 'react'
-import { Cell, Button, Barrage } from '@nutui/nutui-react'
-
-const barrageStyle = {
-  padding: '20px 0',
-  height: '150px',
-  boxSizing: 'border-box',
-}
-const App = () => {
-  const barrageList = [
-    '畫美不看',
-    '不明覺厲',
-    '喜大普奔',
-    '男默女淚',
-    '纍覺不愛',
-    '爺青結-',
-  ]
-  const barrageRef = useRef(null)
-  const addBarrage = () => {
-    const n = Math.random()
-    if (barrageRef.current) {
-      barrageRef.current.add(`隨機——${String(n).substr(2, 10)}`)
-    }
-  }
-  return (
-    <div className="demo">
-      <h2>基礎用法</h2>
-      <Cell className="barrage-demo-wrap" style={barrageStyle}>
-        <Barrage
-          className="barrage-demo"
-          ref={barrageRef}
-          list={barrageList}
-          style={barrageStyle}
-        />
-      </Cell>
-      <div className="test" style={{ textAlign: 'center' }}>
-        <Button type="danger" onClick={addBarrage}>
-          隨機添加
-        </Button>
-      </div>
-    </div>
-  )
-}
-export default App
-```
+<CodeBlock src='h5/demo1.tsx'></CodeBlock>
 
 :::
